[Keymap] Move common code and configuration to userspace for dshields keymaps.
