+++ keyboards/model01/keymaps/dshields/config.h
@@ -1,30 +1,4 @@
 #pragma once
 
-#define RGB_MATRIX_FRAMEBUFFER_EFFECTS
 #define RGB_DIGITAL_RAIN_DROPS    18
-#define USB_MAX_POWER_CONSUMPTION 100
-#define ONESHOT_TAP_TOGGLE        2
-#define ONESHOT_TIMEOUT           3000
-
-#define MOUSEKEY_INTERVAL         20
-#define MOUSEKEY_DELAY            0
-#define MOUSEKEY_TIME_TO_MAX      40
-#define MOUSEKEY_MAX_SPEED        7
-#define MOUSEKEY_WHEEL_DELAY      0
-
-// dynamic macro keys
-#define DM_PLAY DYN_MACRO_PLAY1
-#define DM_STRT DYN_REC_START1
-#define DM_STOP DYN_REC_STOP
-
-// one-shot layer keys
-#define OSL_FUN OSL(FUN)
-
-// one-shot modifier keys
-#define OSMLCTL OSM(MOD_LCTL)
-#define OSMRCTL OSM(MOD_RCTL)
-#define OSMLALT OSM(MOD_LALT)
-#define OSMRALT OSM(MOD_RALT)
-#define OSMLSFT OSM(MOD_LSFT)
-#define OSMRSFT OSM(MOD_RSFT)
