--- keyboards/model01/keymaps/dshields/rules.mk
@@ -1,3 +1,0 @@
-MOUSEKEY_ENABLE = yes
-EXTRAKEY_ENABLE = no
-CONSOLE_ENABLE  = no
