+++ users/dshields/config.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#define RGB_MATRIX_FRAMEBUFFER_EFFECTS
+#define USB_MAX_POWER_CONSUMPTION 100
+#define ONESHOT_TAP_TOGGLE        2
+#define ONESHOT_TIMEOUT           3000
+#define RETRO_TAPPING
+
+#define MOUSEKEY_INTERVAL         20
+#define MOUSEKEY_DELAY            0
+#define MOUSEKEY_TIME_TO_MAX      40
+#define MOUSEKEY_MAX_SPEED        7
+#define MOUSEKEY_WHEEL_DELAY      0
+
